Remove now redundant recipe 2
