+++ chapter-10/README.md
@@ -1,7 +1,6 @@
 # Chapter 10: Writing an Installer
 
 - [Installing your project](recipe-01/README.md)
-- [Installing following GNU standards](recipe-02/README.md)
-- [Generating export headers](recipe-03/README.md)
-- [Exporting your targets](recipe-04/README.md)
-- [Installing a superbuild](recipe-05/README.md)
+- [Generating export headers](recipe-02/README.md)
+- [Exporting your targets](recipe-03/README.md)
+- [Installing a superbuild](recipe-04/README.md)
