+++ chapter-10/recipe-02/cxx-example/src/CMakeLists.txt
@@ -7,6 +7,30 @@ target_sources(message-objs
   )
 set_target_properties(message-objs PROPERTIES
   POSITION_INDEPENDENT_CODE 1
+  CXX_VISIBILITY_PRESET hidden
+  VISIBILITY_INLINES_HIDDEN 1
+  )
+
+include(GenerateExportHeader)
+generate_export_header(message-objs
+  BASE_NAME "message"
+  EXPORT_MACRO_NAME "message_EXPORT"
+  EXPORT_FILE_NAME "messageExport.h"
+  DEPRECATED_MACRO_NAME "message_DEPRECATED"
+  NO_EXPORT_MACRO_NAME "message_NO_EXPORT"
+  STATIC_DEFINE "message_STATIC_DEFINE"
+  NO_DEPRECATED_MACRO_NAME "message_NO_DEPRECATED"
+  DEFINE_NO_DEPRECATED
+  )
+target_include_directories(message-objs
+  PRIVATE
+    ${CMAKE_CURRENT_BINARY_DIR}
+  )
+
+add_library(message-static STATIC $<TARGET_OBJECTS:message-objs>)
+set_target_properties(message-static PROPERTIES
+  COMPILE_DEFINITIONS message_STATIC_DEFINE
+  OUTPUT_NAME "message"
   )
 
 add_library(message-shared SHARED $<TARGET_OBJECTS:message-objs>)
@@ -16,16 +40,14 @@ set_target_properties(message-shared PROPERTIES
   OUTPUT_NAME "message"
   )
 
-add_library(message-static STATIC $<TARGET_OBJECTS:message-objs>)
-set_target_properties(message-static PROPERTIES
-  OUTPUT_NAME "message"
-  )
-
 add_executable(hello-world hello-world.cpp)
-
 target_link_libraries(hello-world
   PUBLIC
     message-shared
+  )
+target_include_directories(hello-world
+  PRIVATE
+    ${CMAKE_CURRENT_BINARY_DIR}
   )
 
 set_target_properties(hello-world
@@ -37,7 +59,7 @@ set_target_properties(hello-world
     INSTALL_RPATH_USE_LINK_PATH ON
   )
 
-install(TARGETS hello-world message-shared message-static
+install(TARGETS hello-world message-static message-shared
   RUNTIME DESTINATION
     ${CMAKE_INSTALL_BINDIR}
   LIBRARY DESTINATION
@@ -46,4 +68,4 @@ install(TARGETS hello-world message-shared message-static
     ${CMAKE_INSTALL_LIBDIR}
   )
 
-install(FILES Message.hpp DESTINATION ${CMAKE_INSTALL_INCLUDEDIR}/message)
+install(FILES ${CMAKE_CURRENT_BINARY_DIR}/messageExport.h Message.hpp DESTINATION ${CMAKE_INSTALL_INCLUDEDIR}/message)
