+++ chapter-10/recipe-02/title.txt
@@ -1,1 +1,1 @@
-Installing following GNU standards
+Generating export headers
