+++ chapter-10/recipe-03/README.md
@@ -1,5 +1,6 @@
-# Generating export headers
+# Exporting your targets
 
 Abstract to be written ...
 
 - [cxx-example](cxx-example/)
+- [cxx-example-3.5](cxx-example-3.5/)
