+++ chapter-10/recipe-03/cxx-example-3.5/src/Message.hpp
