+++ chapter-10/recipe-03/cxx-example/src/Message.cpp
@@ -2,10 +2,22 @@
 
 #include <iostream>
 #include <string>
+#include <uuid/uuid.h>
 
 std::ostream &Message::printObject(std::ostream &os) {
   os << "This is my very nice message: " << std::endl;
-  os << message_;
+  os << message_ << std::endl;
+  os << "...and here is its UUID: " << getUUID();
 
   return os;
+}
+
+std::string getUUID() {
+  uuid_t uuid;
+  uuid_generate(uuid);
+  char uuid_str[37];
+  uuid_unparse_lower(uuid, uuid_str);
+  uuid_clear(uuid);
+  std::string uuid_cxx(uuid_str);
+  return uuid_cxx;
 }
