+++ chapter-10/recipe-03/cxx-example/src/Message.hpp
@@ -17,3 +17,5 @@ private:
   std::string message_;
   std::ostream &printObject(std::ostream &os);
 };
+
+std::string getUUID();
