+++ chapter-10/recipe-03/title.txt
@@ -1,1 +1,1 @@
-Generating export headers
+Exporting your targets
