+++ chapter-10/recipe-04/README.md
@@ -1,6 +1,5 @@
-# Exporting your targets
+# Installing a superbuild
 
 Abstract to be written ...
 
 - [cxx-example](cxx-example/)
-- [cxx-example-3.5](cxx-example-3.5/)
