+++ chapter-10/recipe-04/cxx-example/external/upstream/message/CMakeLists.txt
