--- chapter-10/recipe-04/cxx-example/src/hello-world.cpp
@@ -1,16 +1,0 @@
-#include "Message.hpp"
-
-#include <cstdlib>
-#include <iostream>
-
-int main() {
-  Message say_hello("Hello, CMake World!");
-
-  std::cout << say_hello << std::endl;
-
-  Message say_goodbye("Goodbye, CMake World");
-
-  std::cout << say_goodbye << std::endl;
-
-  return EXIT_SUCCESS;
-}
