+++ chapter-10/recipe-04/cxx-example/src/use_message.cpp
