--- chapter-10/recipe-05/README.md
@@ -1,5 +1,0 @@
-# Installing a superbuild
-
-Abstract to be written ...
-
-- [cxx-example](cxx-example/)
