--- chapter-10/recipe-05/cxx-example/CMakeLists.txt
@@ -1,62 +1,0 @@
-cmake_minimum_required(VERSION 3.6 FATAL_ERROR)
-
-project(recipe-05 LANGUAGES CXX)
-
-set(CMAKE_CXX_STANDARD 11)
-set(CMAKE_CXX_EXTENSIONS OFF)
-set(CMAKE_CXX_STANDARD_REQUIRED ON)
-
-include(GNUInstallDirs)
-
-set(CMAKE_ARCHIVE_OUTPUT_DIRECTORY ${CMAKE_BINARY_DIR}/${CMAKE_INSTALL_LIBDIR})
-set(CMAKE_LIBRARY_OUTPUT_DIRECTORY ${CMAKE_BINARY_DIR}/${CMAKE_INSTALL_LIBDIR})
-set(CMAKE_RUNTIME_OUTPUT_DIRECTORY ${CMAKE_BINARY_DIR}/${CMAKE_INSTALL_BINDIR})
-
-if(CMAKE_INSTALL_PREFIX_INITIALIZED_TO_DEFAULT)
-  set(CMAKE_INSTALL_PREFIX "/usr/local/${PROJECT_NAME}" CACHE PATH "Install path" FORCE)
-endif()
-message(STATUS "${PROJECT_NAME} install: ${CMAKE_INSTALL_PREFIX}")
-
-if(NOT DEFINED CMAKE_BUILD_TYPE OR "${CMAKE_BUILD_TYPE}" STREQUAL "")
-  set(CMAKE_BUILD_TYPE Release CACHE STRING "Build type" FORCE)
-endif()
-message(STATUS "Build type set to ${CMAKE_BUILD_TYPE}")
-
-set(STAGED_INSTALL_PREFIX ${CMAKE_BINARY_DIR}/stage${CMAKE_INSTALL_PREFIX})
-message(STATUS "${PROJECT_NAME} staged install: ${STAGED_INSTALL_PREFIX}")
-
-add_subdirectory(external/upstream)
-
-include(ExternalProject)
-
-ExternalProject_Add(${PROJECT_NAME}_core
-  DEPENDS
-    message_external
-  SOURCE_DIR
-    ${CMAKE_CURRENT_SOURCE_DIR}/src
-  CMAKE_ARGS
-    -DCMAKE_INSTALL_PREFIX=${CMAKE_INSTALL_PREFIX}
-    -DCMAKE_BUILD_TYPE=${CMAKE_BUILD_TYPE}
-    -DCMAKE_CXX_COMPILER=${CMAKE_CXX_COMPILER}
-    -DCMAKE_CXX_FLAGS=${CMAKE_CXX_FLAGS}
-    -DCMAKE_CXX_STANDARD=11
-    -DCMAKE_CXX_EXTENSIONS=OFF
-    -DCMAKE_CXX_STANDARD_REQUIRED=ON
-    -DCMAKE_INSTALL_LIBDIR=${CMAKE_INSTALL_LIBDIR}
-    -DCMAKE_INSTALL_BINDIR=${CMAKE_INSTALL_BINDIR}
-    -DCMAKE_INSTALL_DATADIR=${CMAKE_INSTALL_DATADIR}
-    -DCMAKE_INSTALL_INCLUDEDIR=${CMAKE_INSTALL_INCLUDEDIR}
-    -Dmessage_DIR=${message_DIR}
-  CMAKE_CACHE_ARGS
-    -DCMAKE_PREFIX_PATH:PATH=${CMAKE_PREFIX_PATH}
-  BUILD_ALWAYS
-    1
-  INSTALL_COMMAND
-    DESTDIR=${CMAKE_BINARY_DIR}/stage ${CMAKE_MAKE_PROGRAM} install
-  )
-
-install(
-  DIRECTORY ${STAGED_INSTALL_PREFIX}/
-  DESTINATION ${CMAKE_INSTALL_PREFIX}
-  USE_SOURCE_PERMISSIONS
-  )
