--- chapter-10/recipe-05/cxx-example/src/CMakeLists.txt
@@ -1,39 +1,0 @@
-cmake_minimum_required(VERSION 3.6 FATAL_ERROR)
-
-project(recipe-05_core LANGUAGES CXX)
-
-set(CMAKE_CXX_STANDARD 11)
-set(CMAKE_CXX_EXTENSIONS OFF)
-set(CMAKE_CXX_STANDARD_REQUIRED ON)
-
-include(GNUInstallDirs)
-
-set(CMAKE_ARCHIVE_OUTPUT_DIRECTORY ${CMAKE_BINARY_DIR}/${CMAKE_INSTALL_LIBDIR})
-set(CMAKE_LIBRARY_OUTPUT_DIRECTORY ${CMAKE_BINARY_DIR}/${CMAKE_INSTALL_LIBDIR})
-set(CMAKE_RUNTIME_OUTPUT_DIRECTORY ${CMAKE_BINARY_DIR}/${CMAKE_INSTALL_BINDIR})
-
-find_package(message 1 CONFIG REQUIRED)
-get_property(_loc TARGET message::message PROPERTY LOCATION)
-message(STATUS "Found message: ${_loc} (found version ${message_VERSION})")
-
-add_executable(use_message use_message.cpp)
-target_link_libraries(use_message
-  PUBLIC
-    message::message
-  )
-
-set_target_properties(use_message
-  PROPERTIES
-    MACOSX_RPATH ON
-    SKIP_BUILD_RPATH OFF
-    BUILD_WITH_INSTALL_RPATH OFF
-    INSTALL_RPATH "$ORIGIN/../${CMAKE_INSTALL_LIBDIR}"
-    INSTALL_RPATH_USE_LINK_PATH ON
-  )
-
-install(
-  PROGRAMS
-    ${CMAKE_RUNTIME_OUTPUT_DIRECTORY}/use_message
-  DESTINATION
-    ${CMAKE_INSTALL_PREFIX}/${CMAKE_INSTALL_BINDIR}
-  )
